% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadEncodingTableVN.R
\name{loadEncodingTableVN}
\alias{loadEncodingTableVN}
\title{Load conversion table for Vietnamese characters}
\usage{
loadEncodingTableVN()
}
\value{
data.frame
}
\description{
Load conversion table for Vietnamese characters
}
\details{
The table was adapted from \url{http://vietunicode.sourceforge.net/charset/}
}
